corrections in todo

+++ Todos/assets/js/todo.js
@@ -1,33 +1,40 @@
 let addButton, listItem, todoArray = [], liData, li;
 
 function addTodo() {
-  const today = new Date();
-  const date = `${today.getDate()} - ${(today.toLocaleString('default', { month: 'short' }) )} - ${today.getFullYear()}`;
-  const todoObj = {
-    id: Date.now(),
-    text: document.getElementById('todo').value,
-    date: date
-  };
+  const inputValue = document.getElementById('todo').value;
+  if(inputValue == "") {
+    alert("Enter any text");
+  }
+  else {
+    const today = new Date();
+    const date = `${today.getDate()} - ${(today.toLocaleString('default', { month: 'short' }) )} - ${today.getFullYear()}`;
+    const todoObj = {
+      id: Date.now(),
+      text: inputValue,
+      date: date
+    };
 
-  todoArray.push(todoObj);
-  localStorage.setItem("todos", JSON.stringify(todoArray));
-  displayData(todoObj);
-  reset();
-  // sorting(todoObj);
+    todoArray.push(todoObj);
+    localStorage.setItem("todos", JSON.stringify(todoArray));
+    displayData(todoObj);
+    reset();
+  }
 }
 
 function displayData({id, text, date}) {
-  li = `<li id="${ id }">
+  li = `<li id="${ id }" class = "todo-list">
     <span id="span1">${text}</span>
     ${date}
     <button class="btn btn-info" id="edit" onclick="editTask('${id}', '${text}' ,'${date}');" aria-label="Edit">
       <i class="fas fa-edit" aria-hidden="true"></i>
     </button>
-    <button class="btn btn-danger" id="delete" onclick="deleteTask('${id}');">
-      <i class="fas fa-trash-alt" aria-hidden="true"></i>
+    <button type="button" class="btn btn-danger" data-bs-toggle="modal" data-bs-target="#comfirmBox" onclick="deleteTask('${id}')">
+      <i class="fas fa-trash-alt" aria-hidden="true" ></i>
     </button>
   </li>`;
-
+  // <button class="btn btn-danger" id="delete" data-bs-toggle="modal" data-bs-target="#comfirmBox" onclick="deleteTask('${id}')"  >
+  //     <i class="fas fa-trash-alt" aria-hidden="true"></i>
+  //   </button>
   document.getElementById('task-data').insertAdjacentHTML('beforeend', li);
 }
 
@@ -57,42 +64,25 @@ function editTask(id, text, date) {
   }
 }
 function asce() {
-  debugger
-    todoArray.sort((a, b) => a.id - b.id);
-    //console.log(sorted)
-    console.log(todoArray)
+  const stores_li = document.querySelectorAll('.todo-list');
+  Array.from(stores_li).sort((a, b) => a.id - b.id ).forEach(el => el.parentNode.appendChild(el));
   }
+
 function desc() {
-
-  todoArray.sort((a, b) => b.id - a.id);
-  //console.log(sorted)
-  console.log(todoArray)
-  localStorage.setItem("todos", JSON.stringify(todoArray));
-  todoArray.map(function(ele) {
-    debugger
-    let arrangeData = document.getElementById('task-data').querySelector('li').innerText;
-    console.log(arrangeData)
-    document.querySelector('li').innerText = arrangeData;
-  })
-
+  const stores_li = document.querySelectorAll('.todo-list');
+  Array.from(stores_li).sort((a, b) => b.id - a.id ).forEach(el => el.parentNode.appendChild(el));
 }
-
-
 
 function deleteTask(id) {
-debugger
-  //const deleteId = id;
-  deleteData = todoArray.filter((e) => (e.id !== +id));
-      // todoArray.splice(0,todoArray.findIndex(deleteId));
-  console.log(deleteData)
-  localStorage.setItem("todos", JSON.stringify(deleteData));
+  let confirm = new bootstrap.Modal(document.getElementById('comfirmBox'), {});
+  confirm.show();
+  const confirmValue = document.getElementById('yes').value;
+  console.log(confirmValue)
+  if (confirmValue == "yes") {
+    deleteData = todoArray.filter((e) => (e.id !== +id));
+    document.getElementById(`${id}`).remove();
+    localStorage.setItem("todos", JSON.stringify(deleteData));
+  }
+  confirm.close();
 }
-//
-//   console.log(removedEl);
-//   localStorage.setItem("todos", JSON.stringify(todoArray));
-//   let removeData = todoArray.map(function(element) {
-//     return `${element.text}${element.date}`;
-//   })
-//   document.getElementById("task-data").innerHTML = removeData;
-// }
